perf: remove useless assignments and fix integer overflow
Performance optimizations addressing Copilot review comments:

1. Fix integer overflow in NmfDecompositionTests.cs:109
   - Cast to long before multiplication to prevent overflow on large matrices
   - Changed: matrix.Rows * matrix.Columns
   - To: (long)matrix.Rows * (long)matrix.Columns

2. Remove useless assignments in IcaDecomposition.cs:163-164
   - Removed unused m and n variables in ComputeFastIca
   - Variables were assigned but never read
   - Saves CPU cycles and improves code clarity

3. Remove useless assignments in NmfDecomposition.cs:413-414
   - Removed unused WTW and HHT matrix calculations in Invert()
   - Expensive matrix multiplications that were never used
   - Significant performance improvement for inverse operations

All changes are performance optimizations with no functional changes.

+++ tests/UnitTests/MatrixDecomposition/NmfDecompositionTests.cs
@@ -106,7 +106,7 @@ public class NmfDecompositionTests
             }
         }
 
-        double rmse = Math.Sqrt(totalError / (matrix.Rows * matrix.Columns));
+        double rmse = Math.Sqrt(totalError / ((double)((long)matrix.Rows * (long)matrix.Columns)));
         Assert.True(rmse < 2.0, $"RMSE {rmse} should be less than 2.0");
     }
